[CHORE] 테스트 코드 폴더 구조 변경 (#500)
